requirements.txt
requests
PyPDF2
python-docx
reportlab 